Fix Notifier import cycle
Vitest was not happy with the `require` workaround
