+++ apps/web/src/settings/controllers/NotificationControllers.ts
@@ -7,35 +7,9 @@ SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Com
 Please see LICENSE files in the repository root for full details.
 */
 
-import { logger } from "matrix-js-sdk/src/logger";
-import { PushRuleActionName } from "matrix-js-sdk/src/matrix";
-
 import SettingController from "./SettingController";
-import { MatrixClientPeg } from "../../MatrixClientPeg";
 import { type SettingLevel } from "../SettingLevel";
-
-// .m.rule.master being enabled means all events match that push rule
-// default action on this rule is dont_notify, but it could be something else
-export function isPushNotifyDisabled(): boolean {
-    // Return the value of the master push rule as a default
-    const masterRule = MatrixClientPeg.get()?.pushProcessor.getPushRuleById(".m.rule.master");
-
-    if (!masterRule) {
-        logger.warn("No master push rule! Notifications are disabled for this user.");
-        return true;
-    }
-
-    // If the rule is enabled then check it does not notify on everything
-    return masterRule.enabled && !masterRule.actions.includes(PushRuleActionName.Notify);
-}
-
-function getNotifier(): any {
-    // TODO: [TS] Formal type that doesn't cause a cyclical reference.
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    let Notifier = require("../../Notifier"); // avoids cyclical references
-    if (Notifier.default) Notifier = Notifier.default; // correct for webpack require() weirdness
-    return Notifier;
-}
+import Notifier, { isPushNotifyDisabled } from "../../Notifier";
 
 export class NotificationsEnabledController extends SettingController {
     public getValueOverride(
@@ -44,7 +18,7 @@ export class NotificationsEnabledController extends SettingController {
         calculatedValue: any,
         calculatedAtLevel: SettingLevel | null,
     ): any {
-        if (!getNotifier().isPossible()) return false;
+        if (!Notifier.isPossible()) return false;
 
         if (calculatedValue === null || calculatedAtLevel === "default") {
             return !isPushNotifyDisabled();
@@ -54,15 +28,15 @@ export class NotificationsEnabledController extends SettingController {
     }
 
     public onChange(level: SettingLevel, roomId: string, newValue: any): void {
-        if (getNotifier().supportsDesktopNotifications()) {
-            getNotifier().setEnabled(newValue);
+        if (Notifier.supportsDesktopNotifications()) {
+            Notifier.setEnabled(newValue);
         }
     }
 }
 
 export class NotificationBodyEnabledController extends SettingController {
     public getValueOverride(level: SettingLevel, roomId: string, calculatedValue: any): any {
-        if (!getNotifier().isPossible()) return false;
+        if (!Notifier.isPossible()) return false;
 
         if (calculatedValue === null) {
             return !isPushNotifyDisabled();
